# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SingBetter API",
    "version": "1.0.0",
    "description": "API documentation for website, web app, and PWA flows."
  },
  "servers": [
    { "url": "/api", "description": "Current version" },
    { "url": "/api/v1", "description": "Versioned alias" }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Sign in with email/password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Signed in" },
          "202": { "description": "MFA required" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
...
  }
}